chore(patente): refactor patente components and implement search functionality
